// src/app/shared/models/common-models.ts
export interface Crop {
  name: string;
  crop_type: string;
  id: number;
  max_tree_age: number;
}

export interface State {
  id: number;
  name: string;
  name_en: string;
}

export interface District {
  id: number;
  name: string;
  name_en: string;
}


export interface Taluka {
  id: number;
  name: string;
  name_en: string;
}

export interface Crop {
  id: number;
  name: string;
  name_en: string;
}